==> Odometer Data/Code Scripts/1_Data_Merge.ipynb <==
import pandas as pd
import os


# Specify the folder where your CSV files are located
folder_path = r"E:\IIT Chicago\Sem 4\Data Science Practicum - CSP-572\Datasets\Odometer Data\New folder"

# Get a list of all CSV files in the folder
csv_files = [f for f in os.listdir(folder_path) if f.endswith('.csv')]

# Initialize an empty list to store the DataFrames
dfs = []

# Loop through each CSV file and read it
for file in csv_files:
    file_path = os.path.join(folder_path, file)
    df = pd.read_csv(file_path)
    dfs.append(df)

# Merge all the DataFrames into one
merged_df = pd.concat(dfs, ignore_index=True)

# Save the merged DataFrame to a new CSV file
merged_df.to_csv("E:\IIT Chicago\Sem 4\Data Science Practicum - CSP-572\Datasets\Odometer Data\merged_data.csv", index=False)
