==> Odometer Data/Code Scripts/4_Data_Merge_With_MPG.ipynb <==
import pandas as pd

df1 = pd.read_csv(r"C:\Users\Siddhant\Downloads\merged_all_years.csv")

df2 = pd.read_csv(r"E:\IIT Chicago\Sem 4\Data Science Practicum - CSP-572\Datasets\Odometer Data\mpg_model.csv")

df1.head()

df1.info()

df2 = df2.rename(columns={'Model Year': 'my'})

df2 = df2.rename(columns={'Model': 'model'})

df2.head(50)

# Merge on 'model' and 'my'
df3 = df1.merge(df2, on=['model', 'my'], how='left')

df3.head()

# Data Validation

def filter_data(model_name, my_value):
    """Returns rows matching the given model_api and my values."""
    return df3[(df3["model"] == model_name) & (df3["my"] == my_value)]

model_name_input = "3 Series"  # Replace with actual model_api value
my_input = 2007  # Replace with actual my value

filtered_rows = filter_data(model_name_input, my_input)

filtered_rows.head(20)

models_with_no_data = df3.groupby('model')['Combination MPG'].apply(lambda x: all(x == 'No data found'))

count_models = models_with_no_data.sum()
print(count_models)

count_no_data = (df3['Combination MPG'] == 'No data found').sum()

print(count_no_data)

df3.shape

df3.to_csv('E:\IIT Chicago\Sem 4\Data Science Practicum - CSP-572\Datasets\Odometer Data\complete_data_with_mpg.csv', index=False)
