==> Odometer Data/Code Scripts/5_Fuel_Used_Column_Addition.ipynb <==
import pandas as pd

df1 = pd.read_csv(r"C:\Users\Siddhant\Downloads\New_merged_data.csv")

df2 = pd.read_csv(r"C:\Users\Siddhant\Downloads\Zip+4 - Census Crosswalk\zip_to_blockgrp.csv")

# 1. Explore df1

df1.head()

df1.isnull().sum()

df1.info()

# 2. Create column 'Fuel Used'

df1['Combination MPG'] = pd.to_numeric(df1['Combination MPG'], errors='coerce')

df1['Fuel Used'] = df1['Odometer Difference'] / df1['Combination MPG']

df1[['Combination MPG', 'Odometer Difference', 'Fuel Used']].head()

df1['Fuel Used'].isna().sum()

# 3. Save the file

df1.to_csv('E:\IIT Chicago\Sem 4\Data Science Practicum - CSP-572\Datasets\Odometer Data\complete_dataset.csv', index=False)
