==> Odometer Data/Code Scripts/6_Fuel_Type_Column_Addition.ipynb <==
import pandas as pd

# Original Dataset

df1 = pd.read_csv(r"E:\IIT Chicago\Sem 4\Data Science Practicum - CSP-572\Datasets\Odometer Data\complete_dataset.csv")

df1.head()

df1.info()

# PART 1

# Dataset with model of car and it's fuel type

df2 = pd.read_csv(r'c:\Users\Siddhant\Downloads\VIN_FuelType_Processed_Optimized.csv')

df2.head()

df2.isnull().sum()

df2.info()

df2['Fuel_Type'].value_counts()

# Merge the datasets

df1 = df1.merge(df2[['model', 'Fuel_Type']], on='model', how='left')


df1.head()

df1['Fuel_Type'].unique()


df1['Fuel_Type'].value_counts()

df1 = df1.rename(columns={'Odometer Difference': 'VMT'})

# PART 2

# Add Emission column using calculations

import numpy as np

# Function to calculate emissions based on fuel type and VMT
def calculate_emission(row):
    fuel_type = row['Fuel_Type']
    vmt = row['VMT']
    mpg = row['Combination MPG']

    # Handle cases where VMT or Fuel_Type is NaN
    if pd.isna(vmt) or pd.isna(fuel_type):
        return np.nan

    # Calculate emissions for Gasoline
    if fuel_type == 'Gasoline':
        if pd.isna(mpg):
            return np.nan
        # Gasoline emission calculation
        emission = 8.89e-3 * vmt * (1 / mpg) * (1 / 0.994)

    # Calculate emissions for Diesel
    elif fuel_type == 'Diesel':
        if pd.isna(mpg):
            return np.nan
        # Diesel emission calculation
        emission = 10.180e-3 * vmt * (1 / mpg) * (1 / 0.994)

    # Calculate emissions for Electric
    elif fuel_type == 'Electric':
        # Electric emission calculation
        emission = 823.1 * vmt * (1 / 3.60) * (1 / 1000) * (1 / 2204.6)

    # For other fuel types, return NaN
    else:
        return np.nan

    return emission

# Apply the function to each row and create the 'emissions' column
df1['emissions'] = df1.apply(calculate_emission, axis=1)

df1.head()

df1.isnull().sum()

df1.shape

# Save the dataset

df1.to_csv("E:\IIT Chicago\Sem 4\Data Science Practicum - CSP-572\Datasets\Odometer Data\dataset_emissions_demographic.csv", index=False)
