==> Odometer Data/Code Scripts/7_Blkgrp_And_Demographics_Data_Addition.ipynb <==
import pandas as pd

# Part 1 - Dataset

df1 = pd.read_csv(r"E:\IIT Chicago\Sem 4\Data Science Practicum - CSP-572\Datasets\Odometer Data\complete_data_with_mpg.csv")

df1.head()

df1.info()

df1['zipcode'] = df1['zipcode'].astype('Int64')

df1['zipcode'] = df1['zipcode'].astype('string')

df1.head()

df1.info()

df1 = df1.drop_duplicates(subset='vin', keep='first')

df1.info()

# Part 2 - Original Odometer Dataset Used to get 9 Digit Zipcode

df2 = pd.read_csv(r"E:\IIT Chicago\Sem 4\Data Science Practicum - CSP-572\Datasets\Odometer Data\dataset_emissions_demographic.csv")

# Rename column
df2.rename(columns={'zipcode': 'zip'}, inplace=True)

df2 = df2.merge(df1[['vin', 'zipcode']], on='vin', how='left')

# Part 3 - Add the block group id column to dataset by matching on 9 digit zipcode

df3 = pd.read_csv(r"C:\Users\Siddhant\Downloads\Zip+4 - Census Crosswalk\zip_to_blockgrp.csv")

df3.head()

df3['zipcode'] = df3['zipcode'].astype(str).str.replace('-', '', regex=True)

df3.info()

df3['gid'].nunique()

df3['zipcode'] = df3['zipcode'].astype('Int64')

df3['zipcode'] = df3['zipcode'].astype('string')

df2['zipcode'] = df2['zipcode'].astype('Int64')

df2['zipcode'] = df2['zipcode'].astype('string')

df2 = df2.merge(df3[['blkgrp', 'zipcode']], on='zipcode', how='left')

df2.info()

df2.head()

df2.isnull().sum()

df2['blkgrp'].nunique()

# Part 4 - Add demographics data columns

df4 = pd.read_csv(r"C:\Users\Siddhant\Downloads\CBG_data_acs2022.csv")

df4.head()

df4.info()

df4['geoid'] = df4['geoid'].astype('string')

df4.isnull().sum()

df4.head()

df2['blkgrp'] = df2['blkgrp'].fillna(0).astype('int64')

df2['blkgrp'] = df2['blkgrp'].astype('string')

df2.head()

unique_count = df2['blkgrp'].nunique()

print(f"Number of unique values in 'blkgrp': {unique_count}")

unique_count = df4['geoid'].nunique()

print(f"Number of unique values in 'geoid': {unique_count}")

df_unique = pd.DataFrame(df2['blkgrp'].unique(), columns=['blkgrp'])

common_values_count = df_unique['blkgrp'].isin(df4['geoid']).sum()

print(f"Number of common values: {common_values_count}")

df2 = df2.merge(df4, left_on='blkgrp', right_on='geoid', how='left')

df2.drop('geoid', axis=1, inplace=True)

df2.head()

df2.isnull().sum()

# Save the dataset

df2.to_csv("E:\IIT Chicago\Sem 4\Data Science Practicum - CSP-572\Datasets\Odometer Data\dataset_emissions_demographic_corrected.csv", index=False)
